added role, added restriction on delete api(only admin is allowed), updated sql file

+++ RESTful-in-Peace/src/main/java/com/mohit/example/controller/DoctorRestController.java
@@ -1,5 +1,6 @@
 package com.mohit.example.controller;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -7,6 +8,8 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,7 +21,9 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mohit.example.dto.DoctorInfoDTO;
 import com.mohit.example.model.Doctor;
+import com.mohit.example.model.User;
 import com.mohit.example.service.DoctorService;
+import com.mohit.example.service.UserService;
 
 @RestController
 @RequestMapping("/api")
@@ -26,12 +31,14 @@ public class DoctorRestController {
 
 	@Autowired
 	DoctorService doctorService;
+	@Autowired
+	UserService userService;
 
 	@GetMapping("/doctors")
 	Map<String, Object> getDoctors(@RequestParam Optional<Long> id) {
-		
+
 		Map<String, Object> map = new HashMap<String, Object>();
-		
+
 		try {
 
 			if (id.isPresent()) {
@@ -51,12 +58,12 @@ public class DoctorRestController {
 
 		return map;
 	}
-	
+
 	@PostMapping("/insert/doctor/new")
 	Map<String, Object> insertDoctor(@RequestParam String json) {
-		
+
 		Map<String, Object> map = new HashMap<String, Object>();
-		
+
 		try {
 			DoctorInfoDTO doctorInfoDTO = new ObjectMapper().readValue(json, DoctorInfoDTO.class);
 			Doctor doctor = new Doctor();
@@ -69,12 +76,12 @@ public class DoctorRestController {
 
 		return map;
 	}
-	
+
 	@PutMapping("/update/doctors")
 	Map<String, Object> updateDoctor(@RequestParam Long id, @RequestParam String json) {
-		
+
 		Map<String, Object> map = new HashMap<String, Object>();
-		
+
 		try {
 			DoctorInfoDTO doctorInfoDTO = new ObjectMapper().readValue(json, DoctorInfoDTO.class);
 			Doctor doctor = doctorService.getDoctorById(id);
@@ -87,26 +94,36 @@ public class DoctorRestController {
 
 		return map;
 	}
-	
+
 	@DeleteMapping("/delete/doctors")
-	Map<String, Object> deleteDoctor(@RequestParam Long id) {
-		
+	ResponseEntity<Object> deleteDoctor(Principal principal, @RequestParam Long id) {
+		ResponseEntity<Object> responseEntity;
 		Map<String, Object> map = new HashMap<String, Object>();
-		
+		User user = null;
 		try {
-			Doctor doctor = doctorService.getDoctorById(id);
-			boolean flag = doctorService.deleteDoctor(doctor);
-			if(flag) {
-				map.put("status", "deleted");
+			user = userService.getUserDetails(principal);
+			if (user.getRole().getName().toLowerCase().equals("role_admin")) {
+				Doctor doctor = doctorService.getDoctorById(id);
+				boolean flag = doctorService.deleteDoctor(doctor);
+				if (flag) {
+					map.put("status", "deleted");
+					responseEntity = new ResponseEntity<>(map, HttpStatus.OK);
+				} else {
+					map.put("status", "Please try again");
+					responseEntity = new ResponseEntity<>(map, HttpStatus.INTERNAL_SERVER_ERROR);
+				}
 			}else {
-				map.put("status", "error");
+				map.put("status", "Unauthorized access. Please try again");
+				responseEntity = new ResponseEntity<>(map, HttpStatus.UNAUTHORIZED);
 			}
+
 		} catch (Exception e) {
 			map.put("status", "error");
+			responseEntity = new ResponseEntity<>(map, HttpStatus.INTERNAL_SERVER_ERROR);
 			e.printStackTrace();
 		}
 
-		return map;
+		return responseEntity;
 	}
 
 }
