+++ RESTful-in-Peace/src/main/java/com/mohit/example/controller/PatientRestController.java
@@ -1,5 +1,6 @@
 package com.mohit.example.controller;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -7,6 +8,8 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,7 +21,9 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mohit.example.dto.PatientDTO;
 import com.mohit.example.model.Patient;
+import com.mohit.example.model.User;
 import com.mohit.example.service.PatientService;
+import com.mohit.example.service.UserService;
 
 @RestController
 @RequestMapping("/api")
@@ -26,6 +31,8 @@ public class PatientRestController {
 
 	@Autowired
 	PatientService patientService;
+	@Autowired
+	UserService userService;
 
 	@GetMapping("/patients")
 	Map<String, Object> getDoctors(@RequestParam Optional<Long> id) {
@@ -89,24 +96,35 @@ public class PatientRestController {
 	}
 
 	@DeleteMapping("/delete/patients")
-	Map<String, Object> deleteDoctor(@RequestParam Long id) {
+	ResponseEntity<Object> deleteDoctor(Principal principal, @RequestParam Long id) {
 
+		ResponseEntity<Object> responseEntity;
 		Map<String, Object> map = new HashMap<String, Object>();
+		User user = null;
 
 		try {
-			Patient patient = patientService.getPatientById(id);
-			boolean flag = patientService.deletePatient(patient);
-			if (flag) {
-				map.put("status", "deleted");
+			user = userService.getUserDetails(principal);
+			if (user.getRole().getName().toLowerCase().equals("role_admin")) {
+				Patient patient = patientService.getPatientById(id);
+				boolean flag = patientService.deletePatient(patient);
+				if (flag) {
+					map.put("status", "deleted");
+					responseEntity = new ResponseEntity<>(map, HttpStatus.OK);
+				} else {
+					map.put("status", "Please try again");
+					responseEntity = new ResponseEntity<>(map, HttpStatus.INTERNAL_SERVER_ERROR);
+				}
 			} else {
-				map.put("status", "error");
+				map.put("status", "Unauthorized access. Please try again");
+				responseEntity = new ResponseEntity<>(map, HttpStatus.UNAUTHORIZED);
 			}
 		} catch (Exception e) {
 			map.put("status", "error");
+			responseEntity = new ResponseEntity<>(map, HttpStatus.INTERNAL_SERVER_ERROR);
 			e.printStackTrace();
 		}
 
-		return map;
+		return responseEntity;
 	}
 
 }
